+++ RELEASE_NOTES.md
@@ -19,6 +19,20 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * none
 
+## [0.6.1] - 2023-05-22
+
+### Major Changes
+
+* none
+
+### Minor Changes
+
+* none
+
+### Bugfixes
+
+* Fixed minor logic error in permissions checking, improved testing 
+
 ## [0.6.0] - 2023-05-21
 
 ### Major Changes